+++ app/controllers/api/v1/teacher/medias_controller.rb
@@ -3,13 +3,10 @@ class Api::V1::Teacher::MediasController < Api::V1::TeacherApplicationController
 
   def create
     @media_form = Form::MediaForm.new(params[:media])
-    # TODO: Fix form objects' validation incompatibility
-    # and remove this try-catch
-    begin
-      @media_form.save
+    if @media_form.save
       render json: { id: @media_form.model.id }, status: 200
-    rescue
-      render json: { errors: @media_form.model.errors }, status: 422
+    else
+      render json: { errors: @media_form.errors }, status: 422
     end
   end
